+++ src/client/Player.hpp
@@ -16,7 +16,7 @@ class Player
         virtual std::pair<sf::Int8, sf::Int8> move() {return std::make_pair(0, 0);}
         virtual double rotate([[maybe_unused]] const sf::RenderWindow &window) {return 0.0;}
 
-        void setRotation(int rotation);
+        void setRotation(double rotation);
 
         ~Player() {}
 };
